refactor: move runtime helpers
